Fix the E0 dimension bug, supress WB output
In the measurement routine, the seting of E0 tensor was wrong. Its 3,4
leg is not physical, and has nothing to do with the physical legs of O,
it is the internal leg of MPS, and at the boundary this leg is always 1
dimension. Previously physical dim was set to the 3,4 leg causing
totally wrong measurement in the bulk.

The DMRG output of B-type MPS is turned off. Because the fDMRG is
upgraded to restore the WB recording, this enables relaxation of the
boundary. Because the last site MPS construction requires the info of
updated WB, which is missing in the previous version, leading to the
wrong MPS tensor at the end, and also the wrong free energy at the
boundary. However this upgrading will destroy WB on output, so it no
longer makes sense to output WB from DMRG. Should stick to WA, and make
measurement on it. Due to this convension change, the SET_MPO from
outside is restored.

Add a debug routine to reconstruct many-body wave function from the MPS
tensor.

Update TENSOR.f90 to detect possible leg-rank or dimension miss match
error in TEN_PROD.